refactor(sse): consolidate shopping list SSE into budget SSE
- Remove separate shopping_list_sse.py endpoint
- Add item_* event handlers to budgetSSEClient.js
- Add SAFE_ITEM_FIELDS and broadcast_item_* functions to budget_sse.py
- Update listsManager.js with shopping_list_id filtering (security)
- Remove sseClient.js and per-list SSE indicator from lists.html
- Single SSE connection per user for all budget/shopping events

BREAKING: /api/v1/shopping-lists/{id}/events endpoint removed
Use /api/v1/budget/events for all real-time updates
