+++ backend/app/api/v1/endpoints/shopping_list_items.py
@@ -60,8 +60,8 @@ def _get_sse_broadcast():
     """Lazy import SSE module to avoid circular dependencies."""
     global _sse_module
     if _sse_module is None:
-        from backend.app.api.v1.endpoints import shopping_list_sse
-        _sse_module = shopping_list_sse
+        from backend.app.api.v1.endpoints import budget_sse
+        _sse_module = budget_sse
     return _sse_module
 
 router = APIRouter(
@@ -209,7 +209,6 @@ async def create_shopping_list_item(
     response = ShoppingListItemResponse.model_validate(item)
     sse = _get_sse_broadcast()
     await sse.broadcast_item_created(
-        list_id=item.shopping_list_id,
         item_data=response.model_dump(),
         user_id=current_user.id,
     )
@@ -340,7 +339,6 @@ async def update_shopping_list_item(
     response = ShoppingListItemResponse.model_validate(item)
     sse = _get_sse_broadcast()
     await sse.broadcast_item_updated(
-        list_id=item.shopping_list_id,
         item_data=response.model_dump(),
         user_id=current_user.id,
     )
@@ -403,8 +401,8 @@ async def delete_shopping_list_item(
     # Broadcast SSE event (exclude sender)
     sse = _get_sse_broadcast()
     await sse.broadcast_item_deleted(
-        list_id=list_id,
         item_id=item_id,
+        shopping_list_id=list_id,
         user_id=current_user.id,
     )
 
@@ -476,8 +474,8 @@ async def batch_complete_items(
     for list_id, item_ids in items_by_list.items():
         for item_id in item_ids:
             await sse.broadcast_item_completed(
-                list_id=list_id,
                 item_id=item_id,
+                shopping_list_id=list_id,
                 is_completed=request.is_completed,
                 user_id=current_user.id,
             )
@@ -554,8 +552,8 @@ async def batch_delete_items(
     for list_id, item_ids in items_by_list.items():
         for item_id in item_ids:
             await sse.broadcast_item_deleted(
-                list_id=list_id,
                 item_id=item_id,
+                shopping_list_id=list_id,
                 user_id=current_user.id,
             )
